@@ -16,20 +16,21 @@
     "types:check": "turbo run types:check",
     "version": "changeset version && pnpm install --lockfile-only"
   },
-  "prettier": "@vercel/style-guide/prettier",
   "devDependencies": {
     "@changesets/cli": "^2.27.9",
-    "@typescript-eslint/eslint-plugin": "^8.9.0",
-    "@typescript-eslint/parser": "^8.9.0",
-    "@vercel/style-guide": "^6.0.0",
+    "@eslint/js": "^9.12.0",
+    "@types/eslint__js": "^8.42.3",
     "concurrently": "^9.0.1",
-    "eslint": "^8.57.1",
+    "eslint": "^9.12.0",
+    "eslint-plugin-react": "^7.37.1",
+    "eslint-plugin-react-hooks": "5.1.0-rc-bf7e210c-20241017",
     "eslint-plugin-tailwindcss": "^3.17.5",
     "prettier": "^3.3.3",
     "rimraf": "^6.0.1",
     "tsup": "8.3.0",
     "turbo": "2.1.3",
     "typescript": "^5.6.3",
+    "typescript-eslint": "^8.10.0",
     "vitest": "^2.1.3"
   },
   "packageManager": "pnpm@9.11.0",
