@@ -1,7 +1,5 @@
-import shared from '@vercel/style-guide/prettier';
-
 /** @type {import("prettier").Config} */
 export default {
-  ...shared,
+  singleQuote: true,
   endOfLine: 'lf',
 };
